Fixes #1. Applications are now unique by user ID and/or applicant email.
This commit checks the registrations table to prevent multiple applications from
same user or email address. It also fixes a SQL syntax bug in
thatcamp_registrations_get_registration_by_applicant_email. It also adds a
hidden field so applicant_email is filled out if a user is currently logged in.